# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "3.0版本，基于之前的2.0改造的",
        "title": "用户端接口服务",
        "termsOfService": "http://127.0.0.1/docs/index.html",
        "contact": {
            "name": "追梦小窝",
            "url": "http://github.com/iszmxw",
            "email": "mail@54zm.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0"
    },
    "host": "127.0.0.1",
    "paths": {
        "/v1/api/user/login": {
            "post": {
                "description": "提交注册的邮箱和密码即可登录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录接口"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response._LoginHandler"
                        }
                    }
                }
            }
        },
        "/v1/api/user/send_email_register": {
            "post": {
                "description": "发送注册邮件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "发送注册邮件"
                ],
                "summary": "发送注册邮件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response._OK"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response._LoginHandler": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "reqId": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "properties": {
                        "token": {
                            "description": "登录获取的token",
                            "type": "string"
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response._OK": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "reqId": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}